Adjustments for Multiverse
Added Multiverse proper compatebility;
Zoltan and Lanius now work as intended when patched in Multiverse.
Drones and weapons have been adjusted for ships to use the variety Multiverse offers.
Shipcrew has been adjusted for ships to use the variety Multiverse offers.
Unnecessary code from the vanilla version was purged because it is not needed for Multiverse.